Refactored and organized conveyor commands.
