+++ src/main/java/frc/robot/RobotContainer.java
@@ -14,10 +14,9 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.Subsystem;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
-import edu.wpi.first.wpilibj2.command.button.POVButton;
 import frc.robot.commands.auto.*;
 import frc.robot.commands.intake.AutoIntake;
-import frc.robot.commands.shooter.SingleShot;
+import frc.robot.commands.conveyor.SingleShot;
 import frc.robot.commands.shooter.TeleopShooter;
 import frc.robot.subsystems.*;
 import frc.robot.commands.drivetrain.AlignToTarget;
