+++ src/main/java/frc/robot/commands/intake/ConveyorPeriodic.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.shooter;
+package frc.robot.commands.intake;
 
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
